total mess

+++ src/main/java/org/launchcode/codingevents/controllers/ClubController.java
@@ -26,8 +26,8 @@ public class ClubController {
 	@Autowired
 	ClubAdminRepository clubAdminRepository;
 
-	@GetMapping("/{clubId}")
-	public String getClubDetailPage(Model model, @PathVariable Integer clubId) {
+	@GetMapping()
+	public String getClubDetailPage(Model model, @RequestParam Integer clubId) {
 
 		Optional<Club> result = clubRepository.findById(clubId);
 
@@ -44,8 +44,8 @@ public class ClubController {
 
 	// todo 2: instead of passing in a club object directly to our club admin form, we can set a club on the clubadminDto and then pass that into the form
 	// note: 404 code was written on other branch and this is not doing much validation atm
-	@GetMapping("/{clubId}/admin")
-	public String getAddClubAdminPage(Model model, @PathVariable Integer clubId) {
+	@GetMapping("/admin")
+	public String getAddClubAdminPage(Model model, @RequestParam Integer clubId) {
 
 		Optional<Club> result = clubRepository.findById(clubId);
 		Club club = result.get();
@@ -60,9 +60,10 @@ public class ClubController {
 		return "clubs/createAdmin";
 	}
 
-	@PostMapping("/{clubId}/admin")
-	public String handleClubAdminFormSubmit(@ModelAttribute @Valid ClubClubAdminDTO clubAndAdmin, Errors errors, Model model, @PathVariable String clubId) {
+	@PostMapping("/admin")
+	public String handleClubAdminFormSubmit(@RequestParam Integer clubId, @ModelAttribute @Valid ClubClubAdminDTO clubAndAdmin, Model model, Errors errors) {
 
+		System.out.println("got to the post method");
 		if (errors.hasErrors()) {
 			System.out.println("has errors");
 			System.out.println(errors.getAllErrors());
@@ -73,9 +74,7 @@ public class ClubController {
 //			model.addAttribute("errors", errors);
 //			model.addAttribute("clubAndAdmin", clubAndAdmin);
 
-			String redirectString = "redirect:/clubs/" + clubAndAdmin.getClub().getId() + "/admin";
-
-			return redirectString;
+			return "redirect:/clubs/admin?clubId=" + clubAndAdmin.getClub().getId();
 
 
 		} else {
@@ -86,28 +85,31 @@ public class ClubController {
 			club.setClubAdmin(clubAdmin);
 			clubRepository.save(club);
 
-			return "redirect:/clubs/" + clubAndAdmin.getClub().getId();
+			return "redirect:/clubs?clubId=" + clubAndAdmin.getClub().getId();
 		}
 
 
 	}
 
-	@GetMapping("/{clubId}/admin/delete")
-	public String handleDeleteAdmin(@PathVariable Integer clubId) {
+	@GetMapping("/admin/delete")
+	public String handleDeleteAdmin(@RequestParam() Integer clubId) {
 
 		Optional<Club> result = clubRepository.findById(clubId);
 
 		if (result.isPresent()) {
 			Club club = result.get();
-			Integer adminId = club.getClubAdmin().getId();
-			club.setClubAdmin(null);
+			if (club.getClubAdmin() != null) {
+				Integer adminId = club.getClubAdmin().getId();
+				club.setClubAdmin(null);
 
-			clubAdminRepository.deleteById(adminId);
-			clubRepository.save(club);
+				clubAdminRepository.deleteById(adminId);
+				clubRepository.save(club);
+			}
+
 		}
 
 
-		return "redirect:/clubs/" + clubId;
+		return "redirect:/clubs?clubId=" + clubId;
 	}
 
 }
